SSCS-7090 Move DWP states to common library
