+++ src/main/java/uk/gov/hmcts/reform/sscs/ccd/presubmit/withdrawnappeals/AdminAppealWithdrawnHandler.java
@@ -1,6 +1,6 @@
 package uk.gov.hmcts.reform.sscs.ccd.presubmit.withdrawnappeals;
 
-import static uk.gov.hmcts.reform.sscs.ccd.presubmit.DwpState.WITHDRAWAL_RECEIVED;
+import static uk.gov.hmcts.reform.sscs.ccd.domain.DwpState.WITHDRAWAL_RECEIVED;
 
 import org.springframework.stereotype.Service;
 import uk.gov.hmcts.reform.sscs.ccd.callback.Callback;
