+++ src/main/java/uk/gov/hmcts/reform/sscs/ccd/presubmit/withdrawnappeals/DwpActionWithdrawalHandler.java
@@ -1,7 +1,7 @@
 package uk.gov.hmcts.reform.sscs.ccd.presubmit.withdrawnappeals;
 
-import static uk.gov.hmcts.reform.sscs.ccd.presubmit.DwpState.WITHDRAWAL_RECEIVED;
-import static uk.gov.hmcts.reform.sscs.ccd.presubmit.DwpState.WITHDRAWN;
+import static uk.gov.hmcts.reform.sscs.ccd.domain.DwpState.WITHDRAWAL_RECEIVED;
+import static uk.gov.hmcts.reform.sscs.ccd.domain.DwpState.WITHDRAWN;
 
 import org.springframework.stereotype.Service;
 import uk.gov.hmcts.reform.sscs.ccd.callback.Callback;
