+++ src/test/java/uk/gov/hmcts/reform/sscs/ccd/presubmit/dwpdirectionresponse/DwpDirectionResponseAboutToSubmitHandlerTest.java
@@ -12,11 +12,7 @@ import org.junit.Before;
 import org.junit.Test;
 import uk.gov.hmcts.reform.sscs.ccd.callback.Callback;
 import uk.gov.hmcts.reform.sscs.ccd.callback.PreSubmitCallbackResponse;
-import uk.gov.hmcts.reform.sscs.ccd.domain.CaseDetails;
-import uk.gov.hmcts.reform.sscs.ccd.domain.EventType;
-import uk.gov.hmcts.reform.sscs.ccd.domain.SscsCaseData;
-import uk.gov.hmcts.reform.sscs.ccd.domain.State;
-import uk.gov.hmcts.reform.sscs.ccd.presubmit.DwpState;
+import uk.gov.hmcts.reform.sscs.ccd.domain.*;
 
 public class DwpDirectionResponseAboutToSubmitHandlerTest {
     private static final String USER_AUTHORISATION = "Bearer token";
